refactor header and footer code
