+++ src/components/index.tsx
@@ -6,23 +6,23 @@ export { default as Banner } from './banner';
 export { default as AboutUs } from './aboutUs';
 export { default as BannerLower } from './bannerLower';
 export { default as Partners } from './partners';
-export { default as Footer } from './footer/footer';
+export { default as Footer } from './footer';
 //For cabinet page
 //work space
-export { default as UserWorksSpace } from './forCabinetPage/userWorksSpace/index'
+export { default as UserWorksSpace } from './forCabinetPage/userWorksSpace/index';
 //maps
-export { default as UniversityMaps} from './forCabinetPage/maps/index'
+export { default as UniversityMaps } from './forCabinetPage/maps/index';
 // main options
-export { default as Option1MainMenu } from './forCabinetPage/options/mainOptions/option1/index'
-export { default as Option2MainMenu } from './forCabinetPage/options/mainOptions/option2/index'
-export { default as Option3MainMenu } from './forCabinetPage/options/mainOptions/option3/index'
-export { default as Option4MainMenu } from './forCabinetPage/options/mainOptions/option4/index'
-export { default as Option5MainMenu } from './forCabinetPage/options/mainOptions/option5/index'
+export { default as Option1MainMenu } from './forCabinetPage/options/mainOptions/option1/index';
+export { default as Option2MainMenu } from './forCabinetPage/options/mainOptions/option2/index';
+export { default as Option3MainMenu } from './forCabinetPage/options/mainOptions/option3/index';
+export { default as Option4MainMenu } from './forCabinetPage/options/mainOptions/option4/index';
+export { default as Option5MainMenu } from './forCabinetPage/options/mainOptions/option5/index';
 //user options
-export { default as Option1UserMenu } from './forCabinetPage/options/userOptions/option1/index'
-export { default as Option2UserMenu } from './forCabinetPage/options/userOptions/option2/index'
-export { default as Option3UserMenu } from './forCabinetPage/options/userOptions/option3/index'
-export { default as Option4UserMenu } from './forCabinetPage/options/userOptions/option4/index'
-export { default as Option5UserMenu } from './forCabinetPage/options/userOptions/option5/index'
+export { default as Option1UserMenu } from './forCabinetPage/options/userOptions/option1/index';
+export { default as Option2UserMenu } from './forCabinetPage/options/userOptions/option2/index';
+export { default as Option3UserMenu } from './forCabinetPage/options/userOptions/option3/index';
+export { default as Option4UserMenu } from './forCabinetPage/options/userOptions/option4/index';
+export { default as Option5UserMenu } from './forCabinetPage/options/userOptions/option5/index';
 
 export { default as LoginForm } from './loginForm';
